backend/requirements.txt
flask==2.3.3
psycopg2-binary==2.9.6
flask-cors==4.0.0